feat(infra): alerting, landing and local-ui as single-module Workers
Every Worker whose props are stage-derived is now one module, apps/<app>/src/worker.ts:
the alchemy Worker class the root stack yields, the props Effect over MapleStack, and the
bundle alchemy deploys (main: import.meta.url). The per-app alchemy.run.ts files for the
three are gone, as are their hand-written export default handlers.

alerting registers one Cloudflare.Workers.cron handler per schedule in its init, under
CronEventSourceLive. The source reports every fire as successful, so the platform's retry
never engages, and nothing is lost: the ticks already logged and swallowed their own
failures (catchTickFailure), the schedules re-fire, and the shell logs a failure outside a
tick. The ticks and their routing move to src/scheduled.ts behind a dynamic import, so the
api layer graph stays off the startup path and out of the deploy process, where init also
runs; the test imports that module directly. Each fire builds the layer from the
invocation's env and releases it afterwards, as the ManagedRuntime did.

landing and local-ui yield their Command.Build from the props (dead code in the bundle,
alchemy being sideEffects-free) and serve through Cloudflare.Workers.Request and env.ASSETS,
handing the web Response back via HttpServerResponse.fromWeb. landing's negotiation is a
plain handleRequest in src/handler.ts, which is what its test exercises.

tsconfig.alchemy.json now reaches the api graph through alerting's ticks, so it carries
apps/api's @/ alias and lib list; the CI quality shard installs the three apps for it.
Verified with a dry-run alchemy plan of the whole stack against the dev stage's local
state: every props Effect resolves and alerting's four Cron bindings come from its init.
