+++ .github/workflows/ci.yml
@@ -201,11 +201,11 @@ jobs:
             fail-fast: false
             matrix:
                 include:
-                    # electric-sync too: the root alchemy.run.ts imports its
-                    # single-module Worker, so the Alchemy-entrypoints typecheck
-                    # resolves that graph through apps/electric-sync/node_modules.
+                    # The single-module Workers too: the root alchemy.run.ts imports
+                    # their src/worker.ts, so the Alchemy-entrypoints typecheck
+                    # resolves each graph through that app's node_modules.
                     - shard: quality
-                      install-filters: "@maple/api @maple/electric-sync"
+                      install-filters: "@maple/api @maple/electric-sync @maple/alerting @maple/landing @maple/local-ui"
                     - shard: effect-lint
                       install-filters: ""
                     - shard: build-web
@@ -339,8 +339,9 @@ jobs:
             # every apps/*/alchemy.run.ts) — only the root `typecheck` script chains it,
             # and CI calls turbo directly. Without this they ship unchecked.
             #
-            # That program now reaches electric-sync's single-module Worker, whose
-            # `@maple-dev/effect-sdk` import types from built dist, so the sdk builds first.
+            # That program reaches the single-module Workers (electric-sync, alerting,
+            # landing, local-ui), whose `@maple-dev/effect-sdk` imports type from built
+            # dist, so the sdk builds first.
             - name: Build the libraries the Alchemy entrypoints import
               if: ${{ matrix.shard == 'quality' }}
               run: bun turbo build --filter=@maple-dev/effect-sdk
