+++ apps/alerting/src/worker.ts
@@ -1,557 +1,231 @@
+/**
+ * The alerting Worker in alchemy's single-module form: this file is both the
+ * resource the root stack yields (`yield* Alerting`) and the bundle alchemy
+ * deploys (`main: import.meta.url`). Stage-derived props read `MapleStack`;
+ * `impl` runs once per isolate, on the first event, and registers one handler
+ * per cron. The ticks live in `./scheduled`, imported on the first fire so
+ * the api layer graph stays off the startup path — and out of the deploy
+ * process, where init also runs.
+ *
+ * Alchemy's cron source reports every fire as successful, so the platform's
+ * retry never engages here. Nothing is lost: the ticks already log and
+ * swallow their own failures (`catchTickFailure`), the schedules re-fire on
+ * their own, and a failure outside a tick (the layer build) is logged below.
+ */
 import {
-	ANTICIPATED_ERROR_IDENTIFIERS,
-	AlertDestinationsService,
-	AlertReadModelsService,
-	AlertRuntime,
-	AlertRulesService,
-	AlertsService,
-	AnomalyDetectionService,
-	BucketCacheService,
-	CacheBackendLive,
-	CloudflareAnalyticsService,
-	CloudflareOAuthService,
-	DigestService,
-	EdgeCacheService,
-	EmailService,
-	Env,
-	ErrorActorsService,
-	ErrorIssueReadModelsService,
-	ErrorIssueWorkflowService,
-	ErrorPolicyService,
-	ErrorsService,
-	EscalationService,
-	FixVerificationTickService,
-	HazelOAuthService,
-	layerPg,
-	NotificationDispatcher,
-	IssueFixVerificationService,
-	OrgClickHouseSettingsService,
-	OrgIngestKeysService,
-	OrgMembersService,
-	PlanetScaleOAuthService,
-	PullRequestLookupLive,
-	PlanetScaleService,
-	QueryEngineService,
-	ServiceMapRollupService,
-	TinybirdOrgTokenService,
-	VcsSourceServiceLayer,
-	WarehouseQueryService,
-	summarizeCause,
-	withPgConnectionScope,
-} from "@maple/api/alerting"
-import * as MapleCloudflareSDK from "@maple-dev/effect-sdk/cloudflare"
-import { layerFromEnv, layerFromEnvRecord, runScheduledEffect } from "@maple/effect-cloudflare"
-import { Cause, Effect, Layer, Match } from "effect"
-import type { AlertingWorkerEnv } from "../alchemy.run.ts"
+	CLOUDFLARE_WORKER_PLACEMENT,
+	ManagedMapleDb,
+	MapleStack,
+	type MapleStage,
+	resolveWorkerName,
+} from "@maple/infra/cloudflare"
+import {
+	apnsEnv,
+	appUrlsEnv,
+	authEnv,
+	cloudflareOAuthEnv,
+	ingestKeyCryptoEnv,
+	merge,
+	optionalPlain,
+	optionalSecret,
+	planetScaleOAuthEnv,
+	selfObservabilityEnv,
+	tinybirdEnv,
+} from "@maple/infra/env"
+import * as Cloudflare from "alchemy/Cloudflare"
+import { Cause, Effect, Ref } from "effect"
+import { HttpServerResponse } from "effect/unstable/http"
 
-// Module-scope construction; `flush(env)` resolves env on first call. The
-// in-isolate buffers coalesce concurrent scheduled ticks into one POST per
-// signal.
-const telemetry = MapleCloudflareSDK.make({
-	serviceName: "alerting",
-	serviceNamespace: "core",
-	repositoryUrl: "https://github.com/MapleTechLabs/maple",
-	anticipatedErrorIdentifiers: [...ANTICIPATED_ERROR_IDENTIFIERS],
+/**
+ * The alerting worker's resource bindings, split from the `Config`-sourced env
+ * so `InferEnv` can derive `AlertingWorkerEnv` below.
+ */
+const makeWorkerBindings = ({
+	stage,
+	mapleDb,
+}: {
+	stage: MapleStage
+	mapleDb: Cloudflare.Hyperdrive.Connection | undefined
+}) => ({
+	// Ref stages attach MAPLE_DB via `bindMapleDbRef` in the root stack.
+	...(mapleDb ? { MAPLE_DB: mapleDb } : undefined),
+	// Cross-script binding to the investigation fan-out Workflow hosted by the
+	// api worker. Alert, error, and anomaly ticks start investigations when
+	// incidents open. The first arg is the physical workflow name; `scriptName`
+	// makes this a reference-only binding (the api worker owns the workflow
+	// resource).
+	INVESTIGATION_FANOUT_WORKFLOW: Cloudflare.Workflow<{
+		orgId: string
+		investigationId: string
+		maxWidth: number
+		reservedPasses: number
+		attempt: number
+	}>(resolveWorkerName("investigation-fanout", stage), {
+		className: "InvestigationFanoutWorkflow",
+		scriptName: resolveWorkerName("api", stage),
+	}),
+	// Production only: preview/stg workers run the same email crons against
+	// their own DB branches, so a binding here means every live stage sends
+	// its own copy of onboarding/digest/alert emails to real users.
+	...(stage.kind === "prd"
+		? {
+				EMAIL: Cloudflare.Email.SendEmail("email", {
+					allowedSenderAddresses: ["notifications@noreply.maple.dev"],
+				}),
+			}
+		: undefined),
 })
 
-export const buildLayer = (env: AlertingWorkerEnv) => {
-	// Keep config and binding services on the same invocation-scoped env record;
-	// scheduled handlers already receive the authoritative Cloudflare bindings.
-	const ConfigLive = layerFromEnv(env)
-	const WorkerEnvironmentLive = layerFromEnvRecord(env)
-	const EnvLive = Env.layer.pipe(Layer.provide(ConfigLive))
-
-	const DatabaseLive = layerPg.pipe(Layer.provide(WorkerEnvironmentLive))
-
-	const BaseLive = Layer.mergeAll(EnvLive, DatabaseLive)
-	const AlertRuntimeLive = AlertRuntime.layer
-	const EdgeCacheServiceLive = EdgeCacheService.layer.pipe(Layer.provide(CacheBackendLive))
-
-	const OrgClickHouseSettingsLive = OrgClickHouseSettingsService.layer.pipe(
-		Layer.provide(Layer.mergeAll(BaseLive, EdgeCacheServiceLive)),
-	)
-
-	const TinybirdOrgTokenLive = TinybirdOrgTokenService.layer.pipe(Layer.provide(EnvLive))
-
-	const WarehouseQueryServiceLive = WarehouseQueryService.layer.pipe(
-		Layer.provide(Layer.mergeAll(EnvLive, OrgClickHouseSettingsLive, TinybirdOrgTokenLive)),
-	)
-
-	const BucketCacheServiceLive = BucketCacheService.layer.pipe(Layer.provide(EdgeCacheServiceLive))
-
-	const QueryEngineServiceLive = QueryEngineService.layer.pipe(
-		Layer.provide(WarehouseQueryServiceLive),
-		Layer.provide(EdgeCacheServiceLive),
-		Layer.provide(BucketCacheServiceLive),
-	)
-
-	const HazelOAuthServiceLive = HazelOAuthService.layer.pipe(Layer.provide(BaseLive))
-
-	// EmailService resolves the Cloudflare Email Service `EMAIL` binding from
-	// WorkerEnvironment (delivery binding) in addition to EnvLive (EMAIL_FROM).
-	const EmailServiceLive = EmailService.layer.pipe(
-		Layer.provide(Layer.mergeAll(EnvLive, WorkerEnvironmentLive)),
-	)
-
-	const OrgMembersServiceLive = OrgMembersService.layer.pipe(Layer.provide(EnvLive))
-
-	const AlertDestinationsServiceLive = AlertDestinationsService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(BaseLive, HazelOAuthServiceLive, EmailServiceLive, OrgMembersServiceLive),
-		),
-	)
-
-	const AlertReadModelsServiceLive = AlertReadModelsService.layer.pipe(
-		Layer.provide(Layer.mergeAll(DatabaseLive, WarehouseQueryServiceLive)),
-	)
-
-	const AlertRulesServiceLive = AlertRulesService.layer.pipe(
-		Layer.provide(Layer.mergeAll(DatabaseLive, AlertRuntimeLive)),
-	)
-
-	// WorkerEnvironment is merged in so the incident-open issue-hub hook can see
-	// the cross-script investigation workflow binding. The hoisted AlertRuntime
-	// layer is shared with the narrow rules capability even though the reference
-	// also has production defaults.
-	const AlertsServiceLive = AlertsService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(
-				BaseLive,
-				QueryEngineServiceLive,
-				WarehouseQueryServiceLive,
-				OrgClickHouseSettingsLive,
-				AlertRuntimeLive,
-				HazelOAuthServiceLive,
-				EmailServiceLive,
-				AlertDestinationsServiceLive,
-				AlertReadModelsServiceLive,
-				AlertRulesServiceLive,
-				WorkerEnvironmentLive,
-			),
-		),
-	)
-
-	const NotificationDispatcherLive = NotificationDispatcher.layer.pipe(
-		Layer.provide(Layer.mergeAll(BaseLive, HazelOAuthServiceLive, EmailServiceLive)),
-	)
-
-	const EscalationServiceLive = EscalationService.layer.pipe(
-		Layer.provide(Layer.mergeAll(BaseLive, NotificationDispatcherLive)),
-	)
-
-	const ErrorActorsServiceLive = ErrorActorsService.layer.pipe(Layer.provide(BaseLive))
-	const ErrorIssueWorkflowServiceLive = ErrorIssueWorkflowService.layer.pipe(
-		Layer.provide(Layer.mergeAll(BaseLive, ErrorActorsServiceLive)),
-	)
-	const ErrorPolicyServiceLive = ErrorPolicyService.layer.pipe(Layer.provide(BaseLive))
-	const ErrorIssueReadModelsServiceLive = ErrorIssueReadModelsService.layer.pipe(
-		Layer.provide(Layer.mergeAll(DatabaseLive, WarehouseQueryServiceLive, ErrorIssueWorkflowServiceLive)),
-	)
-
-	// Only reachable from here through an investigation agent's `propose_fix`, but
-	// wired all the same: which worker served the call should not decide whether a
-	// pull-request link arrives with its title and state, or whether attaching an
-	// already-merged PR opens a verification window.
-	const VcsSourceServiceLive = VcsSourceServiceLayer.pipe(Layer.provide(BaseLive))
-
-	const IssueFixVerificationServiceLive = IssueFixVerificationService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(
-				BaseLive,
-				ErrorActorsServiceLive,
-				ErrorIssueWorkflowServiceLive,
-				PullRequestLookupLive.pipe(Layer.provide(VcsSourceServiceLive)),
-			),
-		),
-	)
-
-	// WorkerEnvironment is merged in so incident-open investigations can see the
-	// cross-script fan-out workflow binding.
-	const ErrorsServiceLive = ErrorsService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(
-				BaseLive,
-				WarehouseQueryServiceLive,
-				EdgeCacheServiceLive,
-				NotificationDispatcherLive,
-				ErrorActorsServiceLive,
-				ErrorIssueReadModelsServiceLive,
-				ErrorIssueWorkflowServiceLive,
-				ErrorPolicyServiceLive,
-				IssueFixVerificationServiceLive,
-				WorkerEnvironmentLive,
-			),
-		),
-	)
-
-	// WorkerEnvironment merged in so the tick can reach the fan-out workflow
-	// binding when it opens a verification investigation.
-	const FixVerificationTickServiceLive = FixVerificationTickService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(
-				BaseLive,
-				WarehouseQueryServiceLive,
-				IssueFixVerificationServiceLive,
-				WorkerEnvironmentLive,
-			),
-		),
-	)
-
-	const AnomalyDetectionServiceLive = AnomalyDetectionService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(BaseLive, WarehouseQueryServiceLive, EdgeCacheServiceLive, WorkerEnvironmentLive),
-		),
-	)
-
-	const DigestServiceLive = DigestService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(BaseLive, WarehouseQueryServiceLive, EdgeCacheServiceLive, EmailServiceLive),
-		),
-	)
-
-	const ServiceMapRollupServiceLive = ServiceMapRollupService.layer.pipe(
-		Layer.provide(Layer.mergeAll(BaseLive, WarehouseQueryServiceLive)),
-	)
-
-	const CloudflareOAuthServiceLive = CloudflareOAuthService.layer.pipe(Layer.provide(BaseLive))
-
-	const OrgIngestKeysServiceLive = OrgIngestKeysService.layer.pipe(Layer.provide(BaseLive))
-
-	const CloudflareAnalyticsServiceLive = CloudflareAnalyticsService.layer.pipe(
-		Layer.provide(
-			Layer.mergeAll(
-				BaseLive,
-				WarehouseQueryServiceLive,
-				CloudflareOAuthServiceLive,
-				OrgIngestKeysServiceLive,
-				OrgClickHouseSettingsLive,
-			),
-		),
-	)
-
-	const PlanetScaleOAuthServiceLive = PlanetScaleOAuthService.layer.pipe(Layer.provide(BaseLive))
-
-	const PlanetScaleServiceLive = PlanetScaleService.layer.pipe(
-		Layer.provide(Layer.mergeAll(BaseLive, PlanetScaleOAuthServiceLive)),
-	)
-
-	return Layer.mergeAll(
-		AlertsServiceLive,
-		AnomalyDetectionServiceLive,
-		CloudflareAnalyticsServiceLive,
-		PlanetScaleServiceLive,
-		DigestServiceLive,
-		ErrorsServiceLive,
-		FixVerificationTickServiceLive,
-		EscalationServiceLive,
-		ServiceMapRollupServiceLive,
-		// Exposed in the output, not just provided inward: `withPgConnectionScope`
-		// resolves the `MAPLE_DB` binding from it when it opens the tick's socket.
-		WorkerEnvironmentLive,
-	).pipe(Layer.provideMerge(telemetry.layer), Layer.provideMerge(ConfigLive))
-}
+/**
+ * The alerting worker's runtime env, derived from the declaration above — one
+ * source of truth, imported (type-only) by `./scheduled.ts`.
+ *
+ * `Partial` because a binding's absence is a real runtime state: ref stages
+ * attach MAPLE_DB after the Worker exists, EMAIL is prd-only, and `alchemy
+ * dev` emulation does not cover every binding. The configuration vars stay
+ * `unknown` on purpose: config is read through the Effect ConfigProvider
+ * (`layerFromEnv` → the shared `Env` service), never off `env` directly.
+ */
+export type AlertingWorkerEnv = Partial<Cloudflare.InferEnv<ReturnType<typeof makeWorkerBindings>>> &
+	Record<string, unknown>
 
 /**
- * Standard tick failure isolation. A broken tick must not fail the whole scheduled
- * invocation (several ticks share one cron dispatch), so genuine failures are logged and
- * swallowed — but interrupt-only causes (isolate teardown) are re-raised so they reach
- * `runScheduledEffect`'s `onInterrupt: "graceful"` handling instead of logging a phantom
- * tick failure. Mirrors the per-org guards inside the tick services.
+ * Everything in the alerting worker's env that comes from configuration rather
+ * than from a resource. Largely the api worker's set — the two share 32 keys,
+ * which is why the groups live in `@maple/infra/env`.
  */
-export const catchTickFailure = (label: string) =>
-	Effect.catchCause((cause: Cause.Cause<unknown>) =>
-		Cause.hasInterruptsOnly(cause)
-			? Effect.interrupt
-			: Effect.logError("Alerting tick failed").pipe(
-					Effect.annotateLogs({
-						"error.message": summarizeCause(cause),
-						"maple.alerting.tick": label,
-					}),
-				),
+const configuredEnv = (stage: MapleStage) =>
+	merge(
+		// Alert-rule evaluation runs Tinybird-scoped raw SQL through
+		// TinybirdOrgTokenService, so this is the same set the api worker binds.
+		tinybirdEnv,
+		authEnv,
+		ingestKeyCryptoEnv,
+		appUrlsEnv,
+		// MAPLE_ENDPOINT / MAPLE_ENVIRONMENT / COMMIT_SHA / MAPLE_INGEST_KEY.
+		// MAPLE_ENVIRONMENT is stage-derived and NOT env-overridable: it gates both
+		// the non-prod cron skip below and EmailService.emailAllowed, so an override
+		// would open both at once and leave the prd-only EMAIL binding as the sole
+		// guard.
+		selfObservabilityEnv(stage),
+		// Non-prod stages skip all crons (they share live org data via the prod DB);
+		// set to "1" on a stage to deliberately exercise crons there.
+		optionalPlain("MAPLE_ALERTING_ALLOW_NONPROD"),
+		// Dev-only escape hatch from per-org BYO rows (see apps/api/alchemy.run.ts).
+		optionalPlain("MAPLE_IGNORE_ORG_CLICKHOUSE"),
+		optionalSecret("AUTUMN_SECRET_KEY"),
+		optionalSecret("INTERNAL_SERVICE_TOKEN"),
+		// The alerting worker is where incidents open and resolve, so it is the one
+		// that sends push (platform/Apns.ts) — and it runs the Cloudflare analytics
+		// and PlanetScale inventory pollers, each of which resolves and refreshes
+		// per-org OAuth tokens with the same config the api worker uses.
+		apnsEnv,
+		cloudflareOAuthEnv,
+		planetScaleOAuthEnv,
 	)
 
-interface TickAnnotations {
-	readonly [key: string]: unknown
-}
+/**
+ * Alchemy evaluates a Worker's props wherever the class is yielded — the
+ * deployed bundle included, where they are inert. `__ALCHEMY_RUNTIME__` folds to
+ * `true` there, so the stack-side branch below, and the `@maple/infra` modules
+ * only it reaches, are dead-code-eliminated from what ships.
+ */
+const props = Effect.gen(function* () {
+	if (globalThis.__ALCHEMY_RUNTIME__) return { main: import.meta.url }
+	const { stage, workerDev, devEnv } = yield* MapleStack
+	// Dev stages only; stg/prd bind their own Hyperdrive config by id in the
+	// root stack — `alerting` issues ~97% of the workers' Postgres traffic and
+	// was starving the api's connection pool when the two shared one.
+	const mapleDb = yield* ManagedMapleDb
+	const env = yield* configuredEnv(stage)
+	return {
+		main: import.meta.url,
+		name: resolveWorkerName("alerting", stage),
+		compatibility: { date: "2026-04-08", flags: ["nodejs_compat"] },
+		placement: CLOUDFLARE_WORKER_PLACEMENT,
+		// Under `bun dev`: a sticky port the app's route follows.
+		dev: workerDev("alerting"),
+		workersDev: false,
+		// `devEnv` last, so `.env.local` cannot override the inter-app URLs.
+		env: { ...makeWorkerBindings({ stage, mapleDb }), ...env, ...devEnv },
+	}
+})
 
-const makeTick = <A, E, R>(
-	run: Effect.Effect<A, E, R>,
-	spanKey: string,
-	annotationsFor: (result: A) => TickAnnotations | undefined,
-): Effect.Effect<void, never, R> =>
-	run.pipe(
-		Effect.tap((result) => {
-			const annotations = annotationsFor(result)
-			const namespacedAnnotations =
-				annotations === undefined
-					? undefined
-					: Object.fromEntries(
-							Object.entries(annotations).map(([key, value]) => [
-								`maple.alerting.${key}`,
-								value,
-							]),
+/**
+ * The schedules, each attached to the Worker by its `cron` handler below and
+ * dispatched to a tick group by `selectScheduledProgram`. `0 9 * * *` (the
+ * onboarding drip) was retired when that sequence moved to maple-portal's
+ * campaign system.
+ */
+const ALERTING_CRONS = ["* * * * *", "*/5 * * * *", "*/15 * * * *", "0 * * * *"] as const
+
+/**
+ * Non-prod stages (stg, PR previews) share live org data — stg's Hyperdrive
+ * points at the prod database — so their crons would iterate real orgs with
+ * stage-local Tinybird/Clerk credentials: every tick fails per-org and floods
+ * the error dashboards (and historically sent duplicate emails, see #237).
+ * Same gating philosophy as the prd-only EMAIL binding, with an explicit
+ * override for deliberately exercising crons on a non-prod stage.
+ */
+const cronsEnabled = (env: Record<string, unknown>): boolean =>
+	env.MAPLE_ENVIRONMENT === "production" ||
+	env.MAPLE_ALERTING_ALLOW_NONPROD === "1" ||
+	env.MAPLE_ALERTING_ALLOW_NONPROD === "true"
+
+export default class Alerting extends Cloudflare.Worker<Alerting>()(
+	"alerting",
+	props,
+	Effect.gen(function* () {
+		const exec = yield* Cloudflare.WorkerExecutionContext
+		// Imported on the first fire and kept for the isolate: `./scheduled`
+		// carries the whole api layer graph, which has no business in startup
+		// validation or in the deploy process.
+		const scheduled = yield* Effect.cached(Effect.promise(() => import("./scheduled")))
+		// Once per isolate, not once per fire.
+		const loggedNonProdSkip = yield* Ref.make(false)
+
+		const onFire = (controller: ScheduledController) =>
+			Effect.gen(function* () {
+				const env = yield* Cloudflare.WorkerEnvironment
+				if (!cronsEnabled(env)) {
+					if (!(yield* Ref.getAndSet(loggedNonProdSkip, true))) {
+						yield* Effect.logInfo("Skipping alerting crons on non-production stage").pipe(
+							Effect.annotateLogs({
+								"maple.environment":
+									typeof env.MAPLE_ENVIRONMENT === "string"
+										? env.MAPLE_ENVIRONMENT
+										: "unset",
+								hint: "set MAPLE_ALERTING_ALLOW_NONPROD=1 to run them here",
+							}),
 						)
-			return namespacedAnnotations === undefined
-				? Effect.void
-				: Effect.logInfo("Alerting tick completed").pipe(
-						Effect.annotateLogs({ ...namespacedAnnotations, "maple.alerting.tick": spanKey }),
-					)
-		}),
-		Effect.asVoid,
-		Effect.withSpan(`alerting.${spanKey}_tick`),
-		catchTickFailure(spanKey),
-	)
-
-const alertTick = makeTick(
-	AlertsService.use((alerts) => alerts.runSchedulerTick()),
-	"scheduler",
-	(result) => ({
-		evaluatedCount: result.evaluatedCount,
-		processedCount: result.processedCount,
-		evaluationFailureCount: result.evaluationFailureCount,
-		deliveryFailureCount: result.deliveryFailureCount,
-	}),
-)
-
-const errorTick = makeTick(
-	ErrorsService.use((errors) => errors.runTick()),
-	"error",
-	(result) => ({
-		orgsProcessed: result.orgsProcessed,
-		issuesTouched: result.issuesTouched,
-		incidentsOpened: result.incidentsOpened,
-		incidentsResolved: result.incidentsResolved,
-		issuesReopened: result.issuesReopened,
-		issuesArchived: result.issuesArchived,
-		issuesDeleted: result.issuesDeleted,
-		retentionRan: result.retentionRan,
-	}),
-)
-
-const fixVerificationTick = makeTick(
-	FixVerificationTickService.use((service) => service.runTick()),
-	"fix_verification",
-	(result) => ({
-		examined: result.examined,
-		refuted: result.refuted,
-		investigationsStarted: result.investigationsStarted,
-		verdictsApplied: result.verdictsApplied,
-		skipped: result.skipped,
-		failedRows: result.failedRows,
-	}),
-)
-
-const escalationTick = makeTick(
-	EscalationService.use((escalations) => escalations.runEscalationTick()),
-	"escalation",
-	(result) =>
-		result.processed > 0
-			? {
-					processed: result.processed,
-					sent: result.sent,
-					skipped: result.skipped,
-					failed: result.failed,
-					retried: result.retried,
+					}
+					return
 				}
-			: undefined,
-)
-
-const digestTick = makeTick(
-	DigestService.use((digest) => digest.runDigestTick()),
-	"digest",
-	(result) => ({
-		sentCount: result.sentCount,
-		errorCount: result.errorCount,
-		skipped: result.skipped,
-	}),
-)
-
-// The onboarding drip moved to maple-portal (`camp_onboarding`), which owns the
-// sequence, its send log and its suppression list. `org_onboarding_state` stays
-// here — the in-app checklist still uses the rest of that table — and so do its
-// four `*_email_sent_at` columns, which are what the portal's backfill reads.
-
-const serviceMapRollupTick = makeTick(
-	ServiceMapRollupService.use((rollup) => rollup.runRollupTick()),
-	"service_map_rollup",
-	(result) => ({
-		orgsProcessed: result.orgsProcessed,
-		hoursRolledUp: result.hoursRolledUp,
-		edgesWritten: result.edgesWritten,
-		resolutionsWritten: result.resolutionsWritten,
-		resolutionHoursChecked: result.resolutionHoursChecked,
-		emptyResolutionHours: result.emptyResolutionHours,
-		orgFailures: result.orgFailures,
-	}),
-)
-
-const anomalyTick = makeTick(
-	AnomalyDetectionService.use((anomalies) => anomalies.runTick()),
-	"anomaly",
-	(result) => ({
-		orgsProcessed: result.orgsProcessed,
-		seriesEvaluated: result.seriesEvaluated,
-		incidentsOpened: result.incidentsOpened,
-		incidentsAttached: result.incidentsAttached,
-		incidentsReopened: result.incidentsReopened,
-		incidentsContinued: result.incidentsContinued,
-		incidentsResolved: result.incidentsResolved,
-		orgFailures: result.orgFailures,
-	}),
-)
-
-const cloudflareAnalyticsTick = makeTick(
-	CloudflareAnalyticsService.use((analytics) => analytics.pollAllOrgs()),
-	"cloudflare_analytics",
-	(result) => ({
-		orgs: result.orgs,
-		rowsIngested: result.rowsIngested,
-		skipped: result.skipped,
-		failures: result.failures,
-		perOrg: result.perOrg,
-	}),
-)
-
-const planetScaleTick = makeTick(
-	PlanetScaleService.use((planetscale) => planetscale.pollAllOrgs()),
-	"planetscale",
-	(result) =>
-		result.orgs > 0
-			? {
-					orgs: result.orgs,
-					refreshed: result.refreshed,
-					skipped: result.skipped,
-					failures: result.failures,
-					deployEvents: result.deployEvents,
-				}
-			: undefined,
-)
-
-export interface ScheduledTickPrograms<R = never> {
-	readonly alert: Effect.Effect<void, never, R>
-	readonly anomaly: Effect.Effect<void, never, R>
-	readonly cloudflareAnalytics: Effect.Effect<void, never, R>
-	readonly digest: Effect.Effect<void, never, R>
-	readonly error: Effect.Effect<void, never, R>
-	readonly escalation: Effect.Effect<void, never, R>
-	readonly fixVerification: Effect.Effect<void, never, R>
-	readonly planetScale: Effect.Effect<void, never, R>
-	readonly serviceMapRollup: Effect.Effect<void, never, R>
-}
-
-/**
- * Keep cron routing separate from the Worker shell so the schedule and its
- * concurrency groups can be characterized without acquiring production
- * drivers. The concrete tick Effects remain module-scoped and unchanged.
- */
-export const selectScheduledProgram = <R>(
-	cron: string,
-	ticks: ScheduledTickPrograms<R>,
-): Effect.Effect<void, never, R> =>
-	Match.value(cron).pipe(
-		Match.when("*/5 * * * *", () =>
-			Effect.all([ticks.anomaly, ticks.cloudflareAnalytics, ticks.planetScale], {
-				concurrency: 3,
-				discard: true,
-			}),
-		),
-		Match.when("*/15 * * * *", () => ticks.digest),
-		Match.when("0 * * * *", () => ticks.serviceMapRollup),
-		Match.when("* * * * *", () =>
-			// `fixVerification` is chained onto `error` rather than listed beside it:
-			// a window this minute's error tick just refuted must already be settled
-			// when the verification tick looks at it, and array order under bounded
-			// concurrency does not promise that — alert and escalation finishing
-			// first would have started fixVerification while error still ran.
-			Effect.all([ticks.alert, Effect.andThen(ticks.error, ticks.fixVerification), ticks.escalation], {
-				concurrency: 2,
-				discard: true,
-			}),
-		),
-		// Fail closed: a newly configured cron must not silently inherit the
-		// every-minute alert/error/escalation fan-out.
-		Match.orElse((unknownCron) =>
-			Effect.logWarning("Skipping unknown alerting cron schedule").pipe(
-				Effect.annotateLogs({ cron: unknownCron }),
-			),
-		),
-	)
-
-type ScheduledServices =
-	| AlertsService
-	| AnomalyDetectionService
-	| CloudflareAnalyticsService
-	| DigestService
-	| ErrorsService
-	| EscalationService
-	| FixVerificationTickService
-	| PlanetScaleService
-	| ServiceMapRollupService
-
-const scheduledTicks: ScheduledTickPrograms<ScheduledServices> = {
-	alert: alertTick,
-	anomaly: anomalyTick,
-	cloudflareAnalytics: cloudflareAnalyticsTick,
-	digest: digestTick,
-	error: errorTick,
-	escalation: escalationTick,
-	fixVerification: fixVerificationTick,
-	planetScale: planetScaleTick,
-	serviceMapRollup: serviceMapRollupTick,
-}
-
-interface ScheduledEventLike {
-	readonly cron: string
-}
-
-interface ExecutionContextLike {
-	waitUntil(promise: Promise<unknown>): void
-}
-
-let loggedNonProdSkip = false
-
-export default {
-	async scheduled(
-		event: ScheduledEventLike,
-		env: AlertingWorkerEnv,
-		ctx: ExecutionContextLike,
-	): Promise<void> {
-		// Non-prod stages (stg, PR previews) share live org data — stg's Hyperdrive
-		// points at the prod database — so their crons would iterate real orgs with
-		// stage-local Tinybird/Clerk credentials: every tick fails per-org and floods
-		// the error dashboards (and historically sent duplicate emails, see #237).
-		// Same gating philosophy as the prd-only EMAIL binding, with an explicit
-		// override for deliberately exercising crons on a non-prod stage.
-		const environment = typeof env.MAPLE_ENVIRONMENT === "string" ? env.MAPLE_ENVIRONMENT : ""
-		const allowNonProd =
-			env.MAPLE_ALERTING_ALLOW_NONPROD === "1" || env.MAPLE_ALERTING_ALLOW_NONPROD === "true"
-		if (environment !== "production" && !allowNonProd) {
-			// Once per isolate, not once per tick.
-			if (!loggedNonProdSkip) {
-				loggedNonProdSkip = true
-				console.log(
-					`Skipping alerting crons on non-production stage (MAPLE_ENVIRONMENT=${environment || "unset"}); set MAPLE_ALERTING_ALLOW_NONPROD=1 to run them here`,
+				const { runScheduled, telemetry } = yield* scheduled
+				yield* runScheduled(controller.cron, env).pipe(
+					// Interrupts are isolate teardown: the schedule re-fires anyway, and
+					// they must not be logged as a failed run (same rule as the ticks').
+					Effect.catchCause((cause) =>
+						Cause.hasInterruptsOnly(cause)
+							? Effect.void
+							: Effect.logError("Alerting scheduled run failed", cause).pipe(
+									Effect.annotateLogs({ "maple.alerting.cron": controller.cron }),
+								),
+					),
+					// Drain spans and logs once the fire has settled.
+					Effect.ensuring(exec.waitUntil(Effect.promise(() => telemetry.flush(env)))),
 				)
-			}
-			return
-		}
-		const program = selectScheduledProgram(event.cron, scheduledTicks)
-		try {
-			// Cron ticks cancel gracefully on isolate teardown — the schedule reruns
-			// anyway, and re-raised interrupts (see the per-org catchCause guards in the
-			// tick services) must not surface as failed invocations.
-			// One Postgres socket for the whole tick. Alerting is ~97% of the
-			// workers' Postgres traffic and a single anomaly tick was measured at
-			// 628 dials, each spending one of the Worker's six outbound connection
-			// slots on a handshake; the tick's statements now pipeline over one.
-			await runScheduledEffect(buildLayer(env), withPgConnectionScope(program), ctx, {
-				onInterrupt: "graceful",
 			})
-		} finally {
-			ctx.waitUntil(telemetry.flush(env))
+
+		for (const cron of ALERTING_CRONS) {
+			yield* Cloudflare.Workers.cron(cron, onFire)
 		}
-	},
-	fetch(_request: Request): Response {
-		return new Response("maple-alerting: scheduled only", { status: 404 })
-	},
-}
+
+		return {
+			fetch: Effect.succeed(HttpServerResponse.text("maple-alerting: scheduled only", { status: 404 })),
+		}
+		// The Worker's init IS the entry point: the cron source needs the host Worker,
+		// which only exists here.
+		// oxlint-disable-next-line effecttsgo/strict-effect-provide
+	}).pipe(Effect.provide(Cloudflare.Workers.CronEventSourceLive)),
+) {}
